athena4math/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.11
PyJWT==1.4.0
gunicorn==19.4.3
itsdangerous==0.24
Jinja2==2.7.3
lunardate==0.1.5
MarkupSafe==0.23
pbr==1.8.1
pluggy==0.3.1
py==1.4.31
pyCalverter==1.6.1
pyephem==3.7.6.0
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
stevedore==1.10.0
tox==2.3.1
Werkzeug==0.10.1
wheel==0.26.0
WTForms==2.0.2